requirements.txt
langgraph>=0.2.6
langchain-openai>=0.1.22
langchain>=0.2.14
python-dotenv>=1.0.1
langchain-community>=0.3.15
langchain-postgres>=0.0.12
beautifulsoup4>=4.12.2
pypdf>=5.1.0
psycopg[binary]>=3.2.4
setuptools>=75.8.0
langsmith>=0.3.2
langgraph-checkpoint-sqlite>=2.0.3
duckduckgo-search>=7.3.0
langgraph-cli>=0.1.73
faiss-cpu>=1.7.4
ibapi>=9.81.1
pydantic>=2.6.1
pydantic-settings>=2.1.0
typing-extensions>=4.9.0
